🐛fix : #72 Header 컴포넌트 간소화 및 전역상태관리 코딩 작업
