+++ src/components/MainGnb.tsx
@@ -1,68 +1,45 @@
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import {
-  faHouseChimney,
-  faCar,
-  faRightToBracket,
-  faHeadphones,
-} from '@fortawesome/free-solid-svg-icons';
-import { faFileLines } from '@fortawesome/free-regular-svg-icons';
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
 import { useSession } from '@/hook/session';
 import { signOut } from 'next-auth/react';
 import Image from 'next/image';
+import { useGnbStore } from '@/zustand/useGnbWrap';
 
-export default function MainGnb({ deviceType }: { deviceType: string }) {
+export default function MainGnb({
+  deviceType,
+  modalToggleFn,
+}: {
+  deviceType: string;
+  modalToggleFn?: React.Dispatch<React.SetStateAction<boolean>>;
+}) {
   const session = useSession();
   const pathName = usePathname();
-  let thisPath = pathName.split('/')[1];
+  const { gnb } = useGnbStore();
+
+  // 로그아웃 함수
   const handleSignOut = (e: React.MouseEvent) => {
     e.preventDefault();
     signOut({ callbackUrl: '/' });
   };
-  const gnbArr = [
-    {
-      route: '',
-      text: '홈',
-      icon: faHouseChimney,
-      device: ['mobile'],
-    },
-    {
-      route: 'models',
-      text: '모델',
-      icon: faCar,
-      device: ['web', 'mobile'],
-    },
-    {
-      route: 'drive',
-      text: '전시시승',
-      icon: faRightToBracket,
-      device: ['web', 'mobile'],
-    },
-    {
-      route: 'qna',
-      text: '고객지원',
-      icon: faHeadphones,
-      device: ['web', 'mobile'],
-    },
-    {
-      route: 'info',
-      text: '공지사항',
-      icon: faFileLines,
-      device: ['web', 'mobile'],
-    },
-    {
-      route: 'none',
-      text: '제니시수',
-      icon: faFileLines,
-      device: ['web'],
-    },
-  ];
 
+  // 사이트맵 오픈 함수
+  const handleSiteMapOpen = (e: React.MouseEvent<HTMLElement>): void => {
+    e.preventDefault();
+    if (modalToggleFn !== undefined) {
+      modalToggleFn((prev) => !prev);
+    }
+  };
+
+  // 패스네임 기준 클래스 활성화 함수
+  let thisPath = pathName.split('/')[1];
   const isOnActive = (routeName: string) => (thisPath === routeName ? 'on' : '');
+
+  // MainGnb 컴포넌트의 리턴값
   const handleShowGnb = () => {
+    // 컴포넌트 props가 web일 경우
     if (deviceType === 'web') {
-      const webArr = gnbArr
+      const webArr = gnb
         .filter((item) => item.device.includes(deviceType))
         .map((item) => {
           return (
@@ -91,8 +68,10 @@ export default function MainGnb({ deviceType }: { deviceType: string }) {
           {webArr}
         </ul>
       );
+
+      // 컴포넌트 props가 mobile일 경우
     } else if (deviceType === 'mobile') {
-      const mobileArr = gnbArr
+      const mobileArr = gnb
         .filter((item) => item.device.includes(deviceType))
         .map((item) => {
           return (
@@ -107,6 +86,8 @@ export default function MainGnb({ deviceType }: { deviceType: string }) {
           );
         });
       return <ul className="firstGnb mobileView">{mobileArr}</ul>;
+
+      // 컴포넌트 props가 signSet일 경우
     } else if (deviceType === 'signSet') {
       const signSetFn = () => {
         if (session) {
@@ -130,7 +111,16 @@ export default function MainGnb({ deviceType }: { deviceType: string }) {
         }
       };
       return (
-        <li className="flex items-center justify-end gap-x-[20px] mr-[30px] ">{signSetFn()}</li>
+        <ul className="secondGnb">
+          <li className="flex items-center justify-end gap-x-[20px] mr-[30px] ">{signSetFn()}</li>
+          <li className="sitemapBtn">
+            <div className="p-3 cursor-pointer" onClick={handleSiteMapOpen}>
+              <figure className="relative w-[20px] h-[10px]">
+                <Image src="/images/menu_ham.png" fill sizes="100%" alt="" />
+              </figure>
+            </div>
+          </li>
+        </ul>
       );
     }
   };
