+++ src/components/layout/Header.tsx
@@ -3,17 +3,12 @@
 import { useEffect, useRef, useState } from 'react';
 import { usePathname } from 'next/navigation';
 import Sitemap from './Sitemap';
-import Image from 'next/image';
-import useModalOpenBgFix from '@/hook/useModalOpenBgFix';
 import MainGnb from '../MainGnb';
 
 export default function Header({ isMain }: { isMain: string }) {
   const [modalOn, setModalOn] = useState(false);
   const headerRef = useRef<HTMLHeadElement | null>(null);
   const pathName = usePathname();
-
-  // 모달호출 시 배경 고정 커스텀 훅
-  useModalOpenBgFix(modalOn);
 
   // input 실행시 하단 GNB 사라지게 하는 함수
   useEffect(() => {
@@ -31,12 +26,6 @@ export default function Header({ isMain }: { isMain: string }) {
     return () => window.removeEventListener('resize', thisHeight);
   }, [pathName]);
 
-  // 사이트맵 오픈 함수
-  const handleSiteMapOpen = (e: React.MouseEvent<HTMLElement>): void => {
-    e.preventDefault();
-    setModalOn((prev) => !prev);
-  };
-
   return (
     <header className={isMain} ref={headerRef}>
       <nav className="gnb gnb_web">
@@ -45,16 +34,7 @@ export default function Header({ isMain }: { isMain: string }) {
           <MainGnb deviceType={'mobile'} />
         </div>
         <div className="navWrap">
-          <ul className="secondGnb">
-            <MainGnb deviceType={'signSet'} />
-            <li className="sitemapBtn">
-              <div className="p-3 cursor-pointer" onClick={handleSiteMapOpen}>
-                <figure className="relative w-[20px] h-[10px]">
-                  <Image src="/images/menu_ham.png" fill sizes="100%" alt="" />
-                </figure>
-              </div>
-            </li>
-          </ul>
+          <MainGnb deviceType={'signSet'} modalToggleFn={setModalOn} />
         </div>
       </nav>
